@@ -17,42 +17,33 @@
         "source": "https://github.com/php-tuf/composer-stager"
     },
     "require": {
-        "php": "^7.4 || ^8.0",
-        "symfony/filesystem": "^4.4 || ^5.2 || ^6.1",
-        "symfony/process": "^4.4 || ^5.2 || ^6.1"
+        "php": ">=8.1.0",
+        "ext-json": "*",
+        "symfony/filesystem": "^6.2",
+        "symfony/process": "^6.2"
     },
     "require-dev": {
-        "ext-json": "*",
-        "dealerdirect/phpcodesniffer-composer-installer": "^0.7.2",
-        "ergebnis/composer-normalize": "^2.28",
-        "infection/infection": "^0.26.6",
+        "dealerdirect/phpcodesniffer-composer-installer": "^1.0",
+        "ergebnis/composer-normalize": "^2.30",
         "jangregor/phpstan-prophecy": "^1.0",
-        "pepakriz/phpstan-exception-rules": "^0.12.0",
         "php-parallel-lint/php-parallel-lint": "^1.3",
-        "phpcompatibility/php-compatibility": "^9.3",
         "phpmd/phpmd": "^2.13",
-        "phpmetrics/phpmetrics": "^2.8",
-        "phpro/grumphp-shim": "^1.13",
-        "phpspec/prophecy": "^1.15",
+        "phpro/grumphp-shim": "^1.15",
+        "phpspec/prophecy": "^1.17",
         "phpspec/prophecy-phpunit": "^2.0",
-        "phpstan/extension-installer": "^1.1",
-        "phpstan/phpstan": "^1.8",
-        "phpstan/phpstan-deprecation-rules": "^1.0",
-        "phpstan/phpstan-strict-rules": "^1.4",
-        "phpstan/phpstan-symfony": "^1.2",
-        "phpunit/phpunit": "^9.5",
-        "psalm/phar": "^4.27",
-        "qossmic/deptrac-shim": "^0.24.0",
-        "roave/security-advisories": "dev-latest",
-        "slevomat/coding-standard": "^8.4",
+        "phpstan/extension-installer": "^1.2",
+        "phpstan/phpstan": "^1.10",
+        "phpstan/phpstan-deprecation-rules": "^1.1",
+        "phpstan/phpstan-strict-rules": "^1.5",
+        "phpunit/phpunit": "^9.6",
+        "psalm/phar": "^5.7",
+        "qossmic/deptrac-shim": "^1.0",
+        "slevomat/coding-standard": "^8.8",
         "squizlabs/php_codesniffer": "^3.7",
-        "symfony/config": "^5.4",
-        "symfony/dependency-injection": "^5.4",
-        "symfony/yaml": "^5.4",
+        "symfony/config": "^6.2",
+        "symfony/dependency-injection": "^6.2",
+        "symfony/yaml": "^6.2",
         "thecodingmachine/phpstan-strict-rules": "^1.0"
-    },
-    "suggest": {
-        "roave/security-advisories": "Ensures that your application doesn't have installed dependencies with known security vulnerabilities."
     },
     "minimum-stability": "dev",
     "prefer-stable": true,
