@@ -6,7 +6,6 @@ use PhpTuf\ComposerStager\Domain\Aggregate\PreconditionsTree\CommitterPreconditi
 use PhpTuf\ComposerStager\Domain\Exception\ExceptionInterface;
 use PhpTuf\ComposerStager\Domain\Exception\RuntimeException;
 use PhpTuf\ComposerStager\Domain\Service\FileSyncer\FileSyncerInterface;
-use PhpTuf\ComposerStager\Domain\Service\Precondition\PreconditionInterface;
 use PhpTuf\ComposerStager\Domain\Service\ProcessOutputCallback\ProcessOutputCallbackInterface;
 use PhpTuf\ComposerStager\Domain\Service\ProcessRunner\ProcessRunnerInterface;
 use PhpTuf\ComposerStager\Domain\Value\Path\PathInterface;
@@ -14,14 +13,10 @@ use PhpTuf\ComposerStager\Domain\Value\PathList\PathListInterface;
 
 final class Committer implements CommitterInterface
 {
-    private FileSyncerInterface $fileSyncer;
-
-    private PreconditionInterface $preconditions;
-
-    public function __construct(FileSyncerInterface $fileSyncer, CommitterPreconditionsInterface $preconditions)
-    {
-        $this->fileSyncer = $fileSyncer;
-        $this->preconditions = $preconditions;
+    public function __construct(
+        private readonly FileSyncerInterface $fileSyncer,
+        private readonly CommitterPreconditionsInterface $preconditions,
+    ) {
     }
 
     public function commit(
@@ -29,14 +24,14 @@ final class Committer implements CommitterInterface
         PathInterface $activeDir,
         ?PathListInterface $exclusions = null,
         ?ProcessOutputCallbackInterface $callback = null,
-        ?int $timeout = ProcessRunnerInterface::DEFAULT_TIMEOUT
+        ?int $timeout = ProcessRunnerInterface::DEFAULT_TIMEOUT,
     ): void {
         $this->preconditions->assertIsFulfilled($activeDir, $stagingDir, $exclusions);
 
         try {
             $this->fileSyncer->sync($stagingDir, $activeDir, $exclusions, $callback, $timeout);
         } catch (ExceptionInterface $e) {
-            throw new RuntimeException($e->getMessage(), $e->getCode(), $e);
+            throw new RuntimeException($e->getMessage(), 0, $e);
         }
     }
 }
