@@ -19,6 +19,12 @@ abstract class AbstractPreconditionsTree implements PreconditionsTreeInterface
     /** Gets a status message for when the precondition is unfulfilled. */
     abstract protected function getUnfulfilledStatusMessage(): string;
 
+    /**
+     * The order in which children are evaluated is unspecified and should not be depended upon. There is no
+     * guarantee that the order they are supplied in will have, or continue to have, any determinate effect.
+     *
+     * @see https://github.com/php-tuf/composer-stager/issues/75
+     */
     public function __construct(PreconditionInterface ...$children)
     {
         $this->children = $children;
@@ -27,7 +33,7 @@ abstract class AbstractPreconditionsTree implements PreconditionsTreeInterface
     public function getStatusMessage(
         PathInterface $activeDir,
         PathInterface $stagingDir,
-        ?PathListInterface $exclusions = null
+        ?PathListInterface $exclusions = null,
     ): string {
         return $this->isFulfilled($activeDir, $stagingDir, $exclusions)
             ? $this->getFulfilledStatusMessage()
@@ -37,11 +43,11 @@ abstract class AbstractPreconditionsTree implements PreconditionsTreeInterface
     public function isFulfilled(
         PathInterface $activeDir,
         PathInterface $stagingDir,
-        ?PathListInterface $exclusions = null
+        ?PathListInterface $exclusions = null,
     ): bool {
         try {
             $this->assertIsFulfilled($activeDir, $stagingDir, $exclusions);
-        } catch (PreconditionException $e) {
+        } catch (PreconditionException) {
             return false;
         }
 
@@ -51,7 +57,7 @@ abstract class AbstractPreconditionsTree implements PreconditionsTreeInterface
     public function assertIsFulfilled(
         PathInterface $activeDir,
         PathInterface $stagingDir,
-        ?PathListInterface $exclusions = null
+        ?PathListInterface $exclusions = null,
     ): void {
         foreach ($this->children as $child) {
             $child->assertIsFulfilled($activeDir, $stagingDir, $exclusions);
