@@ -4,15 +4,15 @@ namespace PhpTuf\ComposerStager\Infrastructure\Aggregate\PreconditionsTree;
 
 use PhpTuf\ComposerStager\Domain\Aggregate\PreconditionsTree\CommitterPreconditionsInterface;
 use PhpTuf\ComposerStager\Domain\Aggregate\PreconditionsTree\CommonPreconditionsInterface;
+use PhpTuf\ComposerStager\Domain\Aggregate\PreconditionsTree\NoUnsupportedLinksExistInterface;
 use PhpTuf\ComposerStager\Domain\Aggregate\PreconditionsTree\StagingDirIsReadyInterface;
-use PhpTuf\ComposerStager\Domain\Service\Precondition\CodebaseContainsNoSymlinksInterface;
 
 final class CommitterPreconditions extends AbstractPreconditionsTree implements CommitterPreconditionsInterface
 {
     public function __construct(
-        CodebaseContainsNoSymlinksInterface $codebaseContainsNoSymlinks,
         CommonPreconditionsInterface $commonPreconditions,
-        StagingDirIsReadyInterface $stagingDirIsReady
+        NoUnsupportedLinksExistInterface $noUnsupportedLinksExist,
+        StagingDirIsReadyInterface $stagingDirIsReady,
     ) {
         /** @var array<\PhpTuf\ComposerStager\Domain\Service\Precondition\PreconditionInterface> $children */
         $children = func_get_args();
@@ -32,11 +32,11 @@ final class CommitterPreconditions extends AbstractPreconditionsTree implements 
 
     protected function getFulfilledStatusMessage(): string
     {
-        return 'The preconditions for making staged changes live are fulfilled.'; // @codeCoverageIgnore
+        return 'The preconditions for making staged changes live are fulfilled.';
     }
 
     protected function getUnfulfilledStatusMessage(): string
     {
-        return 'The preconditions for making staged changes live are unfulfilled.'; // @codeCoverageIgnore
+        return 'The preconditions for making staged changes live are unfulfilled.';
     }
 }
