@@ -14,7 +14,7 @@ final class CommonPreconditions extends AbstractPreconditionsTree implements Com
         ActiveAndStagingDirsAreDifferentInterface $activeAndStagingDirsAreDifferent,
         ActiveDirExistsInterface $activeDirExists,
         ActiveDirIsWritableInterface $activeDirIsWritable,
-        ComposerIsAvailableInterface $composerIsAvailable
+        ComposerIsAvailableInterface $composerIsAvailable,
     ) {
         /** @var array<\PhpTuf\ComposerStager\Domain\Service\Precondition\PreconditionInterface> $children */
         $children = func_get_args();
@@ -34,11 +34,11 @@ final class CommonPreconditions extends AbstractPreconditionsTree implements Com
 
     protected function getFulfilledStatusMessage(): string
     {
-        return 'The common preconditions are fulfilled.'; // @codeCoverageIgnore
+        return 'The common preconditions are fulfilled.';
     }
 
     protected function getUnfulfilledStatusMessage(): string
     {
-        return 'The common preconditions are unfulfilled.'; // @codeCoverageIgnore
+        return 'The common preconditions are unfulfilled.';
     }
 }
