@@ -10,7 +10,7 @@ final class StagingDirIsReady extends AbstractPreconditionsTree implements Stagi
 {
     public function __construct(
         StagingDirExistsInterface $stagingDirExists,
-        StagingDirIsWritableInterface $stagingDirIsWritable
+        StagingDirIsWritableInterface $stagingDirIsWritable,
     ) {
         /** @var array<\PhpTuf\ComposerStager\Domain\Service\Precondition\PreconditionInterface> $children */
         $children = func_get_args();
@@ -30,11 +30,11 @@ final class StagingDirIsReady extends AbstractPreconditionsTree implements Stagi
 
     protected function getFulfilledStatusMessage(): string
     {
-        return 'The staging directory is ready to use.'; // @codeCoverageIgnore
+        return 'The staging directory is ready to use.';
     }
 
     protected function getUnfulfilledStatusMessage(): string
     {
-        return 'The staging directory is not ready to use.'; // @codeCoverageIgnore
+        return 'The staging directory is not ready to use.';
     }
 }
