@@ -9,11 +9,8 @@ use PhpTuf\ComposerStager\Domain\Value\PathList\PathListInterface;
 
 final class ActiveDirExists extends AbstractPrecondition implements ActiveDirExistsInterface
 {
-    private FilesystemInterface $filesystem;
-
-    public function __construct(FilesystemInterface $filesystem)
+    public function __construct(private readonly FilesystemInterface $filesystem)
     {
-        $this->filesystem = $filesystem;
     }
 
     public function getName(): string
@@ -29,18 +26,18 @@ final class ActiveDirExists extends AbstractPrecondition implements ActiveDirExi
     public function isFulfilled(
         PathInterface $activeDir,
         PathInterface $stagingDir,
-        ?PathListInterface $exclusions = null
+        ?PathListInterface $exclusions = null,
     ): bool {
         return $this->filesystem->exists($activeDir);
     }
 
     protected function getFulfilledStatusMessage(): string
     {
-        return 'The active directory exists.'; // @codeCoverageIgnore
+        return 'The active directory exists.';
     }
 
     protected function getUnfulfilledStatusMessage(): string
     {
-        return 'The active directory does not exist.'; // @codeCoverageIgnore
+        return 'The active directory does not exist.';
     }
 }
