@@ -9,11 +9,8 @@ use PhpTuf\ComposerStager\Domain\Value\PathList\PathListInterface;
 
 final class StagingDirIsWritable extends AbstractPrecondition implements StagingDirIsWritableInterface
 {
-    private FilesystemInterface $filesystem;
-
-    public function __construct(FilesystemInterface $filesystem)
+    public function __construct(private readonly FilesystemInterface $filesystem)
     {
-        $this->filesystem = $filesystem;
     }
 
     public function getName(): string
@@ -29,18 +26,18 @@ final class StagingDirIsWritable extends AbstractPrecondition implements Staging
     public function isFulfilled(
         PathInterface $activeDir,
         PathInterface $stagingDir,
-        ?PathListInterface $exclusions = null
+        ?PathListInterface $exclusions = null,
     ): bool {
         return $this->filesystem->isWritable($stagingDir);
     }
 
     protected function getFulfilledStatusMessage(): string
     {
-        return 'The staging directory is writable.'; // @codeCoverageIgnore
+        return 'The staging directory is writable.';
     }
 
     protected function getUnfulfilledStatusMessage(): string
     {
-        return 'The staging directory is not writable.'; // @codeCoverageIgnore
+        return 'The staging directory is not writable.';
     }
 }
